+++ src/Ext/Building/Hooks.cpp
@@ -1170,3 +1170,131 @@ DEFINE_HOOK(0x44B6C7, BuildingClass_Mission_Attack_TurretAnim, 0x6)
 
 	return 0;
 }
+
+#pragma region TankBunker
+
+// Jun 23, 2026 - Starkku: Vanilla tank bunker code assumes
+// even-sized foundation. The approach used for even-sized
+// foundations and those that have discrete center cell are
+// mutually exclusive due to pathfinding constraints.
+
+// Handle docking offset calculations.
+DEFINE_HOOK(0x447BE3, BuildingClass_DockingCoord_TankBunker, 0x6)
+{
+	enum { SkipGameCode = 0x447CE1 };
+
+	GET(BuildingClass*, pThis, ESI);
+
+	// Discrete center cell: Docking coord is building center instead of cell closest to approach.
+	if (pThis->Type->GetFoundationWidth() % 2)
+	{
+		const auto coords = pThis->GetCenterCoords();
+
+		// Cleaner hook return at function return is causing problems
+		// due to stack offsets, which is why we're doing it like this.
+		R->ECX(coords.X);
+		R->EDX(coords.Y);
+		R->EAX(&coords);
+
+		return SkipGameCode;
+	}
+
+	return 0;
+}
+
+// Remove now unnecessary (and in fact interfering) rotation code in BuildingClass::UpdateTankBunker().
+DEFINE_HOOK(0x459069, BuildingClass_UpdateTankBunker_CheckOccupants, 0x7)
+{
+	enum { SkipGameCode = 0x4590EF };
+
+	GET(BuildingClass*, pThis, ESI);
+
+	// Discrete center cell: No need to change facing at this stage.
+	if (pThis->Type->GetFoundationWidth() % 2)
+		return SkipGameCode;
+
+	return 0;
+}
+
+// Handle force moving unit to center of bunker.
+DEFINE_HOOK(0x459101, BuildingClass_UpdateTankBunker_RotateToTrack, 0x6)
+{
+	enum { ReturnFromFunction = 0x4591CE };
+
+	GET(BuildingClass*, pThis, ESI);
+	GET(UnitClass*, pUnit, EBP);
+
+	// Discrete center cell: Skip straight to rotating in bunker once finished moving instead of forcing track to center.
+	if (pThis->Type->GetFoundationWidth() % 2)
+	{
+		if (!pUnit->Locomotor->Is_Moving())
+		{
+			pUnit->PrimaryFacing.SetDesired(DirStruct(DirType::South));
+			pThis->TankBunkerState = TankBunkerState::RotateInBunker;
+		}
+
+		return ReturnFromFunction;
+	}
+
+	return 0;
+}
+
+inline void EjectBunkeredUnit(BuildingClass* pThis, UnitClass* pUnit)
+{
+	auto const pType = pUnit->Type;
+	auto const cell = pThis->GetMapCoords() + CellStruct(-1, 1);
+	auto const pNearbyCell = MapClass::Instance.NearByLocation(cell, pType->SpeedType, -1, pType->MovementZone, false, 1, 1, false, false, false, true, cell, false, false);
+	pUnit->SetDestination(MapClass::Instance.GetCellAt(pNearbyCell), false);
+	pUnit->QueueMission(Mission::Move, false);
+}
+
+// Bunker was destroyed, sold or warped away.
+DEFINE_HOOK(0x4593C7, BuildingClass_DestroyTankBunker, 0x6)
+{
+	enum { SkipGameCode = 0x459450 };
+
+	GET(BuildingClass*, pThis, EDI);
+	GET(UnitClass*, pUnit, ESI);
+
+	// Discrete center cell: Send unit out to a nearby cell without forcing drive track.
+	if (pThis->Type->GetFoundationWidth() % 2)
+	{
+		EjectBunkeredUnit(pThis, pUnit);
+		return SkipGameCode;
+	}
+
+	return 0;
+}
+
+// Manual unload of the bunker.
+DEFINE_HOOK(0x4596EC, BuildingClass_UnloadTankBunker, 0x6)
+{
+	enum { SkipGameCode = 0x45980D };
+
+	GET(BuildingClass*, pThis, EDI);
+	GET(UnitClass*, pUnit, ESI);
+
+	// Discrete center cell: Send unit out to a nearby cell without forcing drive track.
+	if (pThis->Type->GetFoundationWidth() % 2)
+	{
+		EjectBunkeredUnit(pThis, pUnit);
+		return SkipGameCode;
+	}
+
+	return 0;
+}
+
+// Update bunker logic every frame if in transition/interacting with unit.
+DEFINE_HOOK(0x44C976, BuildingClass_Mission_Repair_TankBunker, 0x5)
+{
+	GET(BuildingClass*, pThis, EBP);
+
+	auto const pType = pThis->Type;
+
+	if (pType->Bunker && (pThis->TankBunkerState > TankBunkerState::Idle && pThis->TankBunkerState < TankBunkerState::Bunkered))
+		R->EAX(BuildingTypeExt::ExtMap.Find(pType)->BunkerStateUpdateDelay.Get(RulesExt::Global()->BunkerStateUpdateDelay));
+
+	return 0;
+}
+
+#pragma endregion